fix: move page specific components to page directories
