+++ src/app/landing-page/ExploreBlogs.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import Button from './Button';
+import Button from '../../components/Button';
 import { CardStack } from '@/ui/card-stack';
 import Image from 'next/image';
 import Link from 'next/link';
