+++ src/app/landing-page/WhatWeDo.tsx
